pyproject.toml
[project]
name = "foxypack-telegram-telethon"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "shapdi", email = "shapranov.work@gmail.com"},
]
dependencies = ["foxypack>=0.0.11", "foxy-entities>=0.0.3", "telethon>=1.42.0"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "ruff>=0.14.0",
]
test = [
    "pytest>=9.0.2",
    "python-dotenv>=1.0.1"
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
